Code refactoring and additional querying AS for the hostname
Change certificate deployment to be more atomic to avoid the situation
when some of the calls to AS fails and that leave config in
inaproppriate state.
Add querying AS for the configured hostname so in case if the hostname
isn't provided explicitly with `-d` argument the plugin still can try to
retrieve current hostname. It may fail, for instance, hostname
configured in AS can be not supported by Let's Encrypt but it's
additional try to retrieve it anyway.
Rewrite tests to more simple and more resilient to possible changes.
Fix a lot of typos and misspelling in docstrings.